Refactor useWindowDimensions to use context
